# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Zytrixon Mobile API Documentation",
        "description": "API documentation for the Zytrixon Truck Billing mobile application.",
        "contact": {
            "email": "admin@zytrixon.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/trips/{trip_id}/transactions": {
            "post": {
                "tags": [
                    "Trip Transactions"
                ],
                "summary": "Add a transaction to a trip",
                "description": "Creates an expense or recovery transaction for a specific trip, adjusting wallets automatically.",
                "operationId": "2eb73690296e62bec1093d11badd554b",
                "parameters": [
                    {
                        "name": "trip_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "transaction_type",
                                    "expense_category_id",
                                    "amount",
                                    "payment_mode"
                                ],
                                "properties": {
                                    "transaction_type": {
                                        "type": "string",
                                        "enum": [
                                            "expense",
                                            "recovery"
                                        ]
                                    },
                                    "expense_category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 500
                                    },
                                    "payment_mode": {
                                        "type": "string",
                                        "enum": [
                                            "wallet",
                                            "owner_bank",
                                            "fastag"
                                        ]
                                    },
                                    "remarks": {
                                        "type": "string",
                                        "example": "Toll tax"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction added successfully"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transactions/{id}": {
            "delete": {
                "tags": [
                    "Trip Transactions"
                ],
                "summary": "Delete a transaction",
                "description": "Deletes a transaction and reverses its effect on the driver wallet.",
                "operationId": "86ad4e8902634fd885584ee5a0194767",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Super Admin Users"
                ],
                "summary": "Get all users",
                "operationId": "5c2f392a6e65ff2585e59ccce04cb7c2",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{id}/toggle-block": {
            "post": {
                "tags": [
                    "Super Admin Users"
                ],
                "summary": "Block or unblock a user",
                "operationId": "7ed0d499f7a70124408347b77fab2659",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User blocked/unblocked successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{id}/reset-password": {
            "post": {
                "tags": [
                    "Super Admin Users"
                ],
                "summary": "Reset a user password manually",
                "operationId": "04d735b0a80d689c0d1f4f9e9921cc19",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Sanctum token here",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}